refactor: replace leaderboard types to inner folder
