+++ packages/client/src/api/leaderboardAPI/typings.ts
@@ -1,5 +1,5 @@
-import { type LEADERBOARD_TEAM_NAME } from '../config/constants';
-import { type SortOption } from '../store/features/leaderboard/typings';
+import { type LEADERBOARD_TEAM_NAME } from '../../config/constants';
+import { type SortOption } from '../../store/features/leaderboard/typings';
 
 export type LeaderboardRecordData = {
   username: string;
