+++ packages/client/src/store/features/leaderboard/typings.ts
@@ -1,4 +1,4 @@
-import { type LeaderboardRecord } from '../../../api/typings';
+import { type LeaderboardRecord } from '../../../api/leaderboardAPI/typings';
 
 export type SortOption = 'username' | 'score' | 'time' | 'matches' | 'rate';
